requirements.txt
anthropic>=0.40.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
yfinance>=0.2.40
twilio>=9.0.0
ib_insync>=0.9.86
chromadb>=0.5.0
numpy>=1.26.0
pandas>=2.2.0
requests>=2.32.0
python-dotenv>=1.0.0
psutil>=5.9.0
